load HTML content from external file
Writing our step content as a single line of HTML and escaping all the double
quotes would be a huge pain. Much easier to load the content in an external
file.

That also makes it trivial to write the content in anything that can be
converted to HTML -- Markdown, emacs org mode, whatever.