right-clickable links on <li> elements, appear on hover
a proof of concept for putting anchors on individual elements (in
particular, <li> for substeps) and having a right-clickable icon appear
when you hover. This is the standard behavior you see everywhere these
days: for example,

https://github.com/dandrake/workflow-wizard-walkthrough#w3-a-workflow-wizard-walkthrough

The idea is that if a student gets confused on some step, they can send
a URL that links directly to the exact place where they got confused.