fix (?) the platform-specific rendering race condition
I think the root cause here is that we don't have a great way right now
to wait for the DOM to get rendered after setting the inner HTML of the
step content.

But, we were also creating multiple instances of the WorkflowManager and that
made things harder to work with. Now only one, and that seems to address most
of the problems with Safari.

One idea if I need to keep working on this: make a "sentinel" element on
every page, and when we call updatePlatformSpecificElements, check if we
found at least one applicable element. If we get zero, that means the
DOM isn't ready, and we just re-queue ourselves. It's basically a little
hack to implement an event loop; it would be silly, but it would give us
a way to busy wait for the DOM.

Something like

<span id="platformSpecificSentinel class="intel-mac apple-silicon-mac
win other-platform"></span>

at the end of the markup; if that's there, the rest of the DOM should be
in place.